Add improved layout and template for time series charts
