+++ layouts/colours.js
@@ -1,0 +1,7 @@
+module.exports = {
+	remain:'#8897ba', //'#d66d06',
+	leave:'#8eadab', //'#819e9a',//'#458b00',//'#2e6e9e',
+	undecided:'#e9decf',//'#e9decf',//'#cec69b',
+	ftPink:'#fff1e0',
+	font:'#333'
+};
