+++ layouts/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+	singlePoll:require('./single-poll.js'),
+	timeSeries:require('./time-series.js'),
+};
