+++ layouts/single-poll.js
@@ -1,4 +1,3 @@
-
 'use strict'
 
 const d3scale = require('d3-scale');
@@ -16,85 +15,6 @@ const colour = {
 
 const isoShortFormat = d3TimeFormat.format('%Y-%m-%d');
 const ftDateFormat = d3TimeFormat.format('%e %b %Y');
-
-function simpleTimeSeries(width, height, dateDomain, data){
-	let margin = {
-		top:20,
-		left:0,
-		bottom:20,
-		right:0
-	};
-
-	//sort data oldest to newest
-	
-	data = data.sort(function(a,b){
-		return a.startDate.getTime() - b.startDate.getTime();
-	});
-	
-	let smoothedData = []; 
-	for(let i=0; i < data.length; i++){
-		let slice = data.slice(Math.max(0, i-7), i);
-		smoothedData.push({
-			remain: d3Array.mean(slice, (d) => d.remain ),
-			leave: d3Array.mean(slice, (d) => d.leave ),
-			undecided: d3Array.mean(slice, (d) => d.undecided ),
-			date: data[i].startDate
-		});
-	}
-	
-	let filtered = data.filter(function(d,i){
-		let time = d.startDate.getTime();
-		return ( time <= dateDomain[1].getTime() && time >= dateDomain[0].getTime() );
-	});
-	
-	let smoothFiltered = smoothedData.filter(function(d,i){
-		let time = d.date.getTime();
-		return ( time <= dateDomain[1].getTime() && time >= dateDomain[0].getTime() );
-	});
-	
-	let yScale = d3scale.linear()
-		.domain([100,0])
-		.range([0, height - (margin.top + margin.bottom)]);
-		
-	let xScale = d3scale.time()
-		.domain( dateDomain )
-		.range( [0, width - (margin.left + margin.right)] );
-	// start date, end date, two/ three? lines, (final values & positions)
-	let config = {
-		title:'',
-		footer:'Source FT.com',
-		margin:margin,
-		width:width,
-		height:height,
-		fontColour:colour.font,
-		data: filtered,
-		line: {
-			remain: {
-				path:d3Shape.line()
-					.x((d) => xScale(d.date) )
-					.y((d) => yScale(d.remain) )
-					(smoothFiltered),
-				stroke:colour.remain
-			},	
-			leave: {
-				path:d3Shape.line()
-					.x((d) => xScale(d.date) )
-					.y((d) => yScale(d.leave) )
-					(smoothFiltered),
-				stroke:colour.leave
-			},
-			undecided: {
-				path:d3Shape.line()
-					.x((d) => xScale(d.date) )
-					.y((d) => yScale(d.undecided) )
-					(smoothFiltered),
-				stroke:colour.undecided
-			}
-		}
-	};
-	
-	return config;
-}
 
 
 function singlePollLayout(width, height, data, metricEmbed){
@@ -168,7 +88,4 @@ function singlePollLayout(width, height, data, metricEmbed){
 	return config;
 }
 
-module.exports = {
-	singlePoll:singlePollLayout,
-	timeSeries:simpleTimeSeries
-};
+module.exports = singlePollLayout;
